requirements.txt
biopython==1.83
chromadb==0.4.22
sentence-transformers==2.7.0
anthropic==0.28.0
requests==2.31.0
python-dotenv==1.0.0
numpy==1.26.4
